Add files via upload
This is the latest version of the project as of 06/04/2024, and my first upload to github. 

I am doing this for 2 reasons:

1. I want to be able to track my changes in case I break anything and revert back to a healthy state. 
2. Practice with using github for version control.

--- 
Current state:
I have a UI where the user can upload an SRW .zip file and then click buttons that perform the main actions of each individual file (getting log info, checking the licenses, and retrieving the meanings of error codes for the given version of GO-Global). 

Issues:
- My latest attempt was unsuccessful in including the error code descriptions wherever error codes are found. This requires the interleaving of the code in errorCodes.py with the code in ApsLogs.py

TO DO:
- add many more keywords to the log search (e.g. "WERFault.exe")
- Make the whole process automated from one click
- display the information in the main window. Preferably in scrollable windows for the logs and licenses
- display the total seats on support and the total expired seats count from the license check

---